+++ examples/circuits/conditional_if.cpp
@@ -18,7 +18,7 @@ int main() {
     qc.measure(0, 0, false);
 
     // define a boolean predicate of the required form
-    // proxy_to_engine_dits_t -> bool
+    // const_proxy_to_engine_dits_t -> bool
     auto pred = [](auto dits) {
         // returns true when the first dit is 1 at runtime (when run by a
         // quantum engine); in our case, this corresponds to the result of the